Move fips helper into helper/constants outside of sdk
