+++ helper/constants/fips.go
@@ -1,6 +1,6 @@
 // +build !fips_140_3
 
-package consts
+package constants
 
 // IsFIPS returns true if Vault is operating in a FIPS-140-{2,3} mode.
 func IsFIPS() bool {
